requirements.txt
webdavclient==1.0.8
webdavclient3==3.14.6
Pyrogram==2.0.106
colorama==0.4.6
httpx==0.25.2
pillow==11.0.0
TgCrypto==1.2.5
python-telegram-bot==20.6
pytesseract==0.3.13
websocket-client==1.8.0
python-dotenv==1.0.1
tqdm==4.67.1
PyYAML==6.0.2